+++ src/api/nft.ts
@@ -108,6 +108,9 @@ export interface Nft extends BaseNft {
   /** Detailed information on why an NFT was classified as spam. */
   spamInfo?: SpamInfo;
 
-  /** Time at which the NFT was most recently acquired by the user. */
+  /**
+   * Time at which the NFT was most recently acquired by the user. Only
+   * available when specifying `orderBy=transfertime` in the request.
+   */
   acquiredAt?: AcquiredAt;
 }