+++ src/types/types.ts
@@ -1345,8 +1345,9 @@ export enum NftSaleMarketplace {
   SEAPORT = 'seaport',
   LOOKSRARE = 'looksrare',
   X2Y2 = 'x2y2',
-  'WYVERN' = 'wyvern',
-  'CRYPTOPUNKS' = 'cryptopunks',
+  WYVERN = 'wyvern',
+  CRYPTOPUNKS = 'cryptopunks',
+  BLUR = 'blur',
   UNKNOWN = 'unknown'
 }