+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Minor Changes
 
 - Added redundancy to `CoreNamespace.getTokensForOwner()` to handle failures when fetching token metadata.
+- Added support for the `acquiredAt` field for ordered `getNftsForOwner` queries.
 
 ## 2.8.3