+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Added redundancy to `CoreNamespace.getTokensForOwner()` to handle failures when fetching token metadata.
 - Added support for the `acquiredAt` field for ordered `getNftsForOwner` queries. Only available when specifying `orderBy: NftOrdering.TRANSFERTIME` in the request.
+- Added support for the `validAt` response field to `NftNamespace.getNftSales()`.
 
 ## 2.8.3