+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Major Changes
 
+### Minor Changes
+
+## 2.9.0
+
+### Major Changes
+
 - Fixed a bug with in `NftNamespace.getNftSales()` in which there was a typo in `decimal` field of the `NftSaleFeeData` object. The field is now correctly named `decimals`, and `decimal` is marked deprecated to avoid a breaking change.
 
 ### Minor Changes